requirements.txt
absl-py==2.1.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.10
boto3==1.34.90
botocore==1.34.90
Brotli==1.0.9
build==1.2.1
cached-property==1.5.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
conllu==4.4.1
contourpy==1.2.0
cupy==13.1.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.8
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
docopt==0.6.2
en-core-web-trf==3.7.3
entrypoints==0.4
es-core-news-lg==3.7.0
es-dep-news-trf==3.7.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.0
fastjsonschema==2.19.1
fastrlock==0.8.2
ffmpy==0.3.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-pasta==0.2.0
gradio==4.27.0
gradio_client==0.15.1
graphviz==0.20.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.26.0
huggingface_hub==0.22.2
idna==3.4
importlib_metadata==7.1.0
importlib-resources==6.1.1
inflate64==0.3.1
inquirerpy==0.3.4
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.3.2
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lxml==4.9.3
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.0
mistune==3.0.2
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.24.3
opencv-python==4.9.0.80
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pfzy==0.3.4
pickleshare==0.7.5
Pillow==10.1.0
pip==24.0
pip-tools==7.4.1
pipenv==2023.12.1
pipreqs==0.5.0
platformdirs==4.2.1
plotly==5.22.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.42
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
py7zr==0.20.5
pyarrow==14.0.2
pyarrow-hotfix==0.6
pybcj==1.0.1
pycparser==2.22
pycryptodomex==3.15.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.0b4
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.0.9
pyppmd==1.0.0
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
pyzstd==0.15.10
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.10.6
ruff==0.3.5
s3transfer==0.10.1
sacremoses==0.0.43
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.1.99
setuptools==69.2.0
shellingham==1.5.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spaces==0.26.2
spacy==3.7.2
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-streamlit==1.0.6
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.35.0
streamlit-mpl-svg==2.0.1
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.6.4
thinc==8.2.3
threadpoolctl==3.4.0
timm==0.9.10
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
typing-utils==0.1.0
tzdata==2024.1
tzlocal==5.2
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.20.0
validators==0.28.3
virtualenv==20.25.1
wasabi==1.1.2
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
